+++ src/consts.ts
@@ -1,10 +1,10 @@
-export const author_regexp: RegExp = /^:::[\w ,.*_]+:::$/m;
+export const author_regexp: RegExp = /^:::[\w ,.*_'"]+:::$/m;
 export const quote_regexp: RegExp = /^\d+\..+$|^- .+$/m;
 
-export const search_regexp: RegExp = /^(?:[\w,. ]| \|\| )+$/m;
+export const search_regexp: RegExp = /^(?:[\w,. _'"]| \|\| )+$/m;
 
 export const code_block_id_regexp: RegExp = /^id [\w\-_]+$/m;
-export const code_block_search_regexp: RegExp = /^search (?:[\w,. ]| \|\| )+$/m;
+export const code_block_search_regexp: RegExp = /^search (?:[\w,. _'"]| \|\| )+$/m;
 export const code_block_refreshInterval_regexp: RegExp = /^refresh \d+[smhdwMy]$/m;
 export const code_block_customClass_regexp: RegExp = /^customClass \.?[\w\-_]+$/m;